Add test to validate precedence of roll-forward and roll-forward-on-no-candidate-fx settings
The comment on the test documents the desired precedence order.

Moved some of the existing tests which validate the behavior of the two settings together into a new separate test class.
